+++ app/src/main/java/main/WTLibraryApp/IWTLibBookRepositrory.java
@@ -1,4 +1,4 @@
-package books;
+package main.WTLibraryApp;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
