+++ app/src/main/java/main/WTLibraryApp/WTLibraryApp.java
@@ -2,9 +2,10 @@ package main.WTLibraryApp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 
 @SpringBootApplication
-public class WTLibraryApp {
+public class WTLibraryApp extends SpringBootServletInitializer{
 
 	public static void main(String[] args) {
 		SpringApplication.run(WTLibraryApp.class, args);
